// src/shared/enums/user-role.enum.ts
export enum UserRole {
  SUPER_ADMIN = 'SUPER_ADMIN', // 超级管理员
  ADMIN = 'ADMIN', // 管理员
  CONTENT_MANAGER = 'CONTENT_MANAGER', // 内容管理员
  CONTENT_EDITOR = 'CONTENT_EDITOR', // 内容编辑
  MARKETING_MANAGER = 'MARKETING_MANAGER', // 营销经理
  LEGAL_REVIEWER = 'LEGAL_REVIEWER', // 法务审核员
  VIEWER = 'VIEWER', // 查看者
  GUEST = 'GUEST', // 访客
}